refactor: separate textdb.ts and sift.ts
